backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pytz==2023.3.post1
python-dotenv>=1.0.0
